Factor request/response classes into appropriate folders
