+++ lib/reel.rb
@@ -11,12 +11,8 @@ require 'reel/version'
 require 'reel/mixins'
 require 'reel/connection'
 require 'reel/logger'
-require 'reel/request_info'
 require 'reel/request'
-require 'reel/request_body'
-require 'reel/request_parser'
 require 'reel/response'
-require 'reel/response_writer'
 require 'reel/server'
 require 'reel/ssl_server'
 require 'reel/websocket'
