+++ lib/reel/response/writer.rb
