+++ spec/reel/response/writer_spec.rb
@@ -1,8 +1,9 @@
 require 'spec_helper'
 
 describe Reel::Response::Writer do
-  let(:fixture_path) { File.expand_path("../../fixtures/example.txt", __FILE__) }
+  let(:fixture_path) { File.expand_path("../../../fixtures/example.txt", __FILE__) }
   let(:expected_response) { "HTTP/1.1 200 OK\r\nContent-Length: 56\r\n\r\n#{File.read(fixture_path)}" }
+
   it "streams static files" do
     with_socket_pair do |socket, peer|
       writer = described_class.new(socket)
